+++ web/api/auth_handler.go
@@ -6,6 +6,7 @@ import (
 	"net/http"
 
 	"github.com/bakku/easyalert"
+	"github.com/bakku/easyalert/random"
 )
 
 // AuthHandler accepts a JSON object containing email and password.
@@ -52,7 +53,7 @@ func (h AuthHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 			return
 		}
 
-		writeError(w, http.StatusInternalServerError, "An unknown error occured.")
+		writeError(w, http.StatusInternalServerError, "an unknown error occured")
 		return
 	}
 
@@ -62,6 +63,65 @@ func (h AuthHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 	}
 
 	responseBody := authResponseBody{user.Token}
+
+	responseBodyBytes, err := json.Marshal(responseBody)
+	if err != nil {
+		writeError(w, http.StatusInternalServerError, "could not marshal response body")
+		return
+	}
+
+	body, err := prettifyJSON(string(responseBodyBytes))
+	if err != nil {
+		writeError(w, http.StatusInternalServerError, "could not prettify json response")
+		return
+	}
+
+	w.Header().Set("Content-Type", "application/json; charset=UTF-8")
+	w.WriteHeader(http.StatusOK)
+	w.Write([]byte(body))
+}
+
+// AuthRefreshHandler requires the token of the user in the Authorization
+// header and refreshes and returns a new token for the user.
+type AuthRefreshHandler struct {
+	UserRepo easyalert.UserRepository
+}
+
+// ServeHTTP handles the HTTP request.
+func (h AuthRefreshHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
+	token, ok := getUserToken(r)
+	if !ok {
+		writeError(w, http.StatusUnauthorized, "Missing or invalid Authorization header.")
+		return
+	}
+
+	user, err := h.UserRepo.FindUser("WHERE token = $1", token)
+	if err != nil {
+		if err == easyalert.ErrRecordDoesNotExist {
+			writeError(w, http.StatusUnauthorized, "Invalid token.")
+			return
+		}
+
+		writeError(w, http.StatusInternalServerError, "an unknown error occured")
+		return
+	}
+
+	newToken, err := random.String(easyalert.UserTokenLength)
+	if err != nil {
+		writeError(w, http.StatusInternalServerError, "could not generate token")
+		return
+	}
+
+	user.Token = newToken
+
+	user, err = h.UserRepo.UpdateUser(user)
+	if err != nil {
+		writeError(w, http.StatusInternalServerError, "could not update token")
+		return
+	}
+
+	var responseBody authResponseBody
+	responseBody.Token = user.Token
 
 	responseBodyBytes, err := json.Marshal(responseBody)
 	if err != nil {